Baking art
The Bellow is what you're using the fan the flames in the heat up section of the minigame, when that part of the minigame is done make the bellow dissapear.
For the Flames, I'd like it if the framerate was low (like 6 fps? whatever makes it look good)
Added a still frame for the first and last stage for the bread (Dough and Burnt), in case its needed, for example having the dough lie there unchaning before the baking part of the minigame starts.